Add routines for geometrical search
The commit adds codes to create STRtree indices for geometrical search.
A function to find an element (or elements) from a position is added.
Note that the return is just 0-based element indices in the tree
indices, not actual element indices yet.
This commit also fix a bug in read_hgrid_gr3.

SCH-568