Add a method to read vgrid.in
Instead of reading zCoordinates and calculating the layer thickness from
them, it is possible to use the sigma ratios to calculate the layer
thickness. The vertical integration can use the ratios directly without
calculating the layer thickness at every node every time step.
To do so, `vgrid.in` can be read in for later use.
Note that the code is implemented only for `ivcor = 1`, not for the sz
grid. The implementation uses Pandas read_fwf, but it is still slow
because of very long lines in `vgrid.in`.

SCH-585